crypto/ayyMessage/requirements.txt
bcrypt==3.1.7
capstone==4.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
filelock==3.0.12
idna==2.8
importlib-metadata==1.5.0
intervaltree==3.0.2
Mako==1.1.1
MarkupSafe==1.1.1
packaging==20.1
paramiko==2.7.1
parse==1.14.0
pluggy==0.13.1
psutil==5.6.7
pwntools==4.0.1
py==1.8.1
pycparser==2.19
pycryptodome==3.9.6
pyelftools==0.26
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.6
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.22.0
ROPGadget==6.0
six==1.14.0
sortedcontainers==2.1.0
toml==0.10.0
tox==3.14.3
unicorn==1.0.2rc1
urllib3==1.25.8
virtualenv==16.7.9
zipp==2.1.0
